requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
